backend/requirements.txt
# backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
piexif==1.1.3
pytesseract==0.3.10
imagehash==4.3.1
numpy==1.26.2
scipy==1.11.4
aiohttp==3.9.1
pydantic==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0